+++ packages/docs/sidebars.js
@@ -38,7 +38,7 @@ const sidebars = {
         "main/custom-stages",
         "main/publishing-to-the-workshop",
         "main/gotchas",
-        "main/isaacscript-common-in-lua",
+        "main/isaacscript-in-lua",
       ],
     },
     "main/change-log",