+++ packages/docs/docs/main/change-log.md
@@ -11,7 +11,7 @@ This page lists the changes to the IsaacScript framework.
 
 ## September 2nd, 2022 (Unreleased)
 
-- `isaacscript-common` can now be imported by Lua mods. See [the documentation](https://isaacscript.github.io/main/isaacscript-common-in-lua).
+- `isaac-typescript-definitions` and `isaacscript-common` can now be imported by Lua mods. See [the documentation](https://isaacscript.github.io/main/isaacscript-in-lua).
 - `isaacscript init` now has a `--dev` flag for setting up a mod that will be testing out a development version of `isaacscript-common`. For more information, see the [README](https://github.com/IsaacScript/isaacscript#working-with-isaacscript-common).
 - IsaacScript now provides the `isaac-lua-polyfill` package, which assists in unit testing your mods using e.g. Jest. Thanks to Aleksander Ciesielski for this.
 - Added the following helper functions: